Renamed User Entity and updated Entity itself

+++ src/main/java/com/js/calendar/controller/UserController.java
@@ -26,7 +26,7 @@ public class UserController {
     }
 
     @GetMapping("/{userId}")
-    public ResponseEntity<User> getUserById(@PathVariable("userId") Integer userId) {
+    public ResponseEntity<User> getUserById(@PathVariable("userId") Long userId) {
         Optional<User> user = userService.getUser(userId);
 
         return user.map(value -> new ResponseEntity<>(value, HttpStatus.OK)).orElseGet(() -> new ResponseEntity<>(new User(), HttpStatus.NOT_FOUND));
@@ -39,13 +39,13 @@ public class UserController {
     }
 
     @PutMapping("/{userId}")
-    public ResponseEntity updateUser(@PathVariable("userId") Integer userId, @RequestBody User user) {
+    public ResponseEntity updateUser(@PathVariable("userId") Long userId, @RequestBody User user) {
         userService.updateUser(userId, user);
         return new ResponseEntity(HttpStatus.OK);
     }
 
     @DeleteMapping("/{userId}")
-    public ResponseEntity deleteUser(@PathVariable("userId") Integer userId) {
+    public ResponseEntity deleteUser(@PathVariable("userId") Long userId) {
         userService.deleteUser(userId);
         return new ResponseEntity(HttpStatus.NO_CONTENT);
     }
