+++ src/main/java/com/js/calendar/service/UserService.java
@@ -7,7 +7,7 @@ import java.util.Optional;
 public interface UserService {
     Iterable<User> getUsers();
     void addUser(User user);
-    void updateUser(Integer id, User user);
-    Optional<User> getUser(Integer id);
-    void deleteUser(Integer id);
+    void updateUser(Long id, User user);
+    Optional<User> getUser(Long id);
+    void deleteUser(Long id);
 }
